player points -> double

+++ Player.java
@@ -5,7 +5,7 @@ public class Player implements Comparator<Player>{
   String nationality;
   int age;
   int rating;
-  int points;
+  double points;
 
   public Player(String n, String na, int a, int rating){
     this.name = n;
